+++ eslint.config.mjs
@@ -1,0 +1,48 @@
+import path from 'node:path'
+import { fileURLToPath } from 'node:url'
+import tsParser from '@typescript-eslint/parser'
+import js from '@eslint/js'
+import { FlatCompat } from '@eslint/eslintrc'
+import { defineConfig } from 'eslint/config'
+import prettier from 'eslint-plugin-prettier'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
+  allConfig: js.configs.all,
+})
+
+export default defineConfig([
+  {
+    ignores: [
+      '**/node_modules',
+      '**/dist',
+      '**/docs',
+      '**/eslint.config.mjs',
+      '**/jest.config.ts',
+    ],
+  },
+  {
+    extends: compat.extends(
+      'eslint:recommended',
+      'plugin:@typescript-eslint/recommended',
+      'prettier',
+    ),
+
+    plugins: {
+      prettier,
+    },
+
+    languageOptions: {
+      parser: tsParser,
+      ecmaVersion: 2022,
+      sourceType: 'script',
+
+      parserOptions: {
+        project: './tsconfig.eslint.json',
+      },
+    },
+  },
+])