+++ jest.config.ts
@@ -1,0 +1,12 @@
+import type { Config } from 'jest'
+import { createDefaultEsmPreset } from 'ts-jest'
+
+const presetConfig = createDefaultEsmPreset()
+
+export default {
+  ...presetConfig,
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+    '#(.*)': '<rootDir>/node_modules/$1',
+  },
+} satisfies Config